# ENH: Decrease wall time of `ma.cov` and `ma.corrcoef`