--- a/doc/release/upcoming_changes/26285.performance.rst
+++ b/doc/release/upcoming_changes/26285.performance.rst
@@ -1,5 +1,5 @@
 ``ma.cov`` and ``ma.corrcoef`` are now significantly faster
 -----------------------------------------------------------
-The private function `ma.extras._covhelper` has been refactored along with 
-`ma.cov` and `ma.corrcoef`. They are now significantly faster, particularly 
-on large, masked arrays.
+The private function has been refactored along with `ma.cov` and 
+`ma.corrcoef`. They are now significantly faster, particularly on large, 
+masked arrays.