@christopher-titchen 2024-05-23
> The benchmarks should be added to the benchmark suite if there are no benchmarks currently for ma.cov and ma.corrcoef

I have added benchmarks for `ma.cov` and `ma.corrcoef` in `bench_ma` as requested.

| Change   | Before [e1bf1d63] <main>   | After [a23e2664] <fix-ma-cov>   |   Ratio | Benchmark (Parameter)          |
|----------|----------------------------|---------------------------------|---------|--------------------------------|
| -        | 1.30±0.1ms                 | 625±30μs                        |    0.48 | bench_ma.Cov.time_cov('large') |

| Change   | Before [e1bf1d63] <main>   | After [a23e2664] <fix-ma-cov>   |   Ratio | Benchmark (Parameter)                    |
|----------|----------------------------|---------------------------------|---------|------------------------------------------|
| -        | 17.6±1ms                   | 315±4μs                         |    0.02 | bench_ma.Corrcoef.time_corrcoef('small') |
| -        | 2.03±0.03s                 | 817±40μs                        |    0    | bench_ma.Corrcoef.time_corrcoef('large') |