@christopher-titchen 2024-05-23
> Is this well tested? If there are doubts, maybe the astropy test suite has some tests for this?

The changes are well tested. `ma.cov` produces exactly the same result as before, just significantly faster. `ma.corrcoef` produces a slightly different result in cases where the observations between a pair of variables are not aligned, and the same result as before in all other cases. This is intended and the reasons for this are outlined above and in the release notes.

The refactored functions pass all of the previous tests and I have added a new test for `ma.extras._covhelper` to reflect the changes.