@mattip 2024-05-24
Lets put this in. The massive performance improvement for `ma.corrcoef` seems worth the small change in output.