@christopher-titchen 2024-05-29
> Lets put this in. The massive performance improvement for `ma.corrcoef` seems worth the small change in output.

Thank you for taking the time to review and approve my PR. I can see that one check for the Accelerate LP64 and ILP64 interfaces on macOS ARM64 is failing, but it seems unrelated to the PR and was successful on my previous commit, after which I only updated the documentation.

https://github.com/numpy/numpy/actions/runs/9214520468/job/25350855711?pr=26285#step:7:143

Apologies, but as this is my first contribution here, I am unsure about the approval process. Is there anything else that I need to do to wrap up the PR?